chore: simplify and remove unused code

+++ app/src/main/java/com/github/libretube/ui/fragments/PlayerFragment.kt
@@ -359,7 +359,7 @@ class PlayerFragment : Fragment(), OnlinePlayerOptions {
             }
         }
 
-        binding.playerMotionLayout.progress = 1.toFloat()
+        binding.playerMotionLayout.progress = 1F
         binding.playerMotionLayout.transitionToStart()
 
         val activity = requireActivity()
